Neighbourhood Segmentation and Clustering
In this assignment, we will Explore, Segment, and Cluster the neighborhoods in the city of Toronto. For the Toronto neighborhood data, a Wikipedia page exists that has all the information we need to explore and cluster the neighborhoods in Toronto.

We need to Scrape the Wikipedia page and Wrangle the data, Clean it, and then read it into a pandas dataframe so that it is in a structured format.